+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `mcmeta.py 2.0.0`: Rewritten to instead read and process an input template.
 - `error.py 1.1.0`: Generally improved logic and log handling.
 - `arguments.py 1.0.1`: Remove unnecessary `get_args` helper function.
+- `arguments 1.1.0`: Improve help message.
  
 ### Fixed